Change mode_t from unsigned long to unsigned int.
The mode_t type really should be an unsigned int, since there's no use in
it expanding on 64-bit systems. Gettext was having compilation problems
with the unsigned long version due to printf format warnings.

+++ apps/libc/include/sys/types.h
@@ -103,7 +103,7 @@ typedef unsigned long long ino_t;
 // Define a type used for file mode bits.
 //
 
-typedef unsigned long mode_t;
+typedef unsigned int mode_t;
 
 //
 // Define a type used for link count (both hard and symbolic).
